boot: rename DebugDumpBootvars->DumpBootvars and move info boot/debug.go
